vhost: Fix unused_assignments warning
This test changes flags and bits to make the object invalid, and checks
that `is_valid()` catches it.  It always reverts that change after every
test.

As a result, there is an assignment at the end to make the object valid
again, with no check remaining afterwards.  For me, the compiler
complains about that, so add a final assertion to make it happy.

Signed-off-by: Hanna Czenczek <hreitz@redhat.com>